Move image format class to a separate file
